refactor(example): use barrel import for GoogleBooksService
