+++ projects/example-app/src/app/books/guards/book-exists.guard.ts
@@ -4,7 +4,7 @@ import { select, Store } from '@ngrx/store';
 import { Observable, of } from 'rxjs';
 import { catchError, filter, map, switchMap, take, tap } from 'rxjs/operators';
 
-import { GoogleBooksService } from '@example-app/core/services/google-books.service';
+import { GoogleBooksService } from '@example-app/core/services';
 import { BookActions } from '@example-app/books/actions';
 import * as fromBooks from '@example-app/books/reducers';
